Major Refactoring complete.  Read Comment
Right so changelog is in the root of the project folder along with a
flowchart showing the flow of control or whatever they call it.

Added many many managers for many many things in an attempt to make our
skeleton game more modular and easier to code.  The current game's state
has pretty much removed the texture switching (sorry) but the code is
all there still, just commented out ready for the person who did it to
reactivate it in the new system.  We have 20 enemies spawning at all
four corners and following the player.  Enjoy people and give feedback
if you want.